Move JobSubmitter, JobType, StorageClient to api.submission
The submission API (JobSubmitter trait, JobType, JobSubmitterConstants,
StorageClient) has no Spark dependencies — only ai.chronon.api.* and
slf4j. Moving to api/ makes it available to all modules without
pulling in Spark.

Backwards compatibility: spark.submission package object re-exports
all types so existing imports continue to work. No consumer changes
needed.

Verified: api, spark, cloud_aws, cloud_gcp all compile.
Tests: JobSubmitterTest (6/6), KyuubiSubmitterTest (14/14) pass.
